chore: Create a module named messenger. This module will be able to inform a user of how many references of a given value they are using. The main objective of this module is to limit how many times a value is referenced.
Implement Logger: a trait which implements the following three functions:
fn warning(&self, msg: &str);
fn info(&self, msg: &str);
fn error(&self, msg: &str);

Implement the Tracker structure with the following fields:
logger: a reference to Logger.
value: the count of how many times the value was referenced. It should not exceed max.
max: the max count of references.

Add the following associated functions to Tracker:
new: that initializes the structure.
set_value: that sets the value. It should compare the number of references to value and max to work out the percentage used. It should write to the following traits if it exceeds the specified usage percentage:
percentage >= 100%: 'Error: you are over your quota!' should be written to error.
percentage >= 70% and percentage < 100%: 'Warning: you have used up over X% of your quota! Proceeds with precaution' should be written to warning, where X should be replaced with the calculated percentage.
peek: that will take a peek at how much usage the variable already has. It should write 'Info: you are using up to X% of your quota' to the info trait function. X should be replaced with the calculated percentage.