chore: Define the following structures:
Workers: containing:
drops: that will save the number of dropped threads.
states: that will save the state of multiple threads. If the thread is not dropped, the state will be false, and will be true otherwise.
Thread: containing:
pid: the id of the thread.
cmd: the name of the thread.
parent: a link to the structure Workers. (Tip: this should be a reference).

You'll need to also add the following associated functions to the structures:
Workers :
new: that creates a default worker.
new_worker: that returns a tuple with the pid and a new Thread. This function must receive a String representing the cmd.
is_dropped: that receives a pid and returns a bool that indicates the state of the thread.
track_worker: which returns a usize representing the length of the states vector. (The index of the next new thread).
add_drop: which is called by the Drop trait. It will receive a pid that will be used to change the state of the thread. If the state of that thread is true then it will panic with the message 'X is already dropped', where X represents the pid). Otherwise it should change the state to true and increment the drops field by 1.

Thread:
new_thread: that initializes a new thread.
skill: that drops the thread.