chore: Fruit: increases the strength by 4 units for each kilogram of fruit consumed. Meat: has a weight in kilograms, and it's pure fat content as a decimal fraction. The remaining weight of meat which is not pure fat is considered to be protein. Each kilogram of protein increases the strength by 4 units. Each kilogram of fat increases the strength by 9 units.
Define the Food trait for Fruit and Meat. The required method gives returns the amount of strength that the food provides.

Implement the std::fmt::Display trait for the Player structure, so that when {} corresponds to a Player insider a println macro, it will print 3 lines:
First: the name of the player.
Second: strength, score and the money.
Third: The player's list of weapons.