chore: Define the structure ThreeDVector, that represents a 3 dimensional vector.
In physics, these vectors are represented as ai, bj and ck. a, b and c are real numbers. i, j and k represent the direction in the Cartesian plane, along the axises x, y and z respectively. Therefore, we use the fields i, j and k in the structure.

Take a look how the operations Addition and Subtraction work for a 3 dimensional vector, and implement them by creating the std::ops::Add and std::ops::Sub traits.